Move the initialization of the HRE to its own module
