+++ v-next/hardhat/src/cli.ts
@@ -1,9 +1,5 @@
 import { register } from "tsx/esm/api";
 
-// Note: We import the builtin plugins' types here, so that any type extension
-// they may have gets loaded.
-import "./internal/builtin-plugins/index.js";
-
 // We enable the sourcemaps before loading main, so that everything except this
 // small file is loaded with sourcemaps enabled.
 process.setSourceMapsEnabled(true);
