+++ v-next/hardhat/test/internal/builtin-plugins/clean/task-action.ts
@@ -13,9 +13,9 @@ import {
 } from "@ignored/hardhat-vnext-utils/fs";
 import { useFixtureProject } from "@nomicfoundation/hardhat-test-utils";
 
-import { createHardhatRuntimeEnvironment } from "../../../../src/hre.js";
 import cleanAction from "../../../../src/internal/builtin-plugins/clean/task-action.js";
 import { getCacheDir } from "../../../../src/internal/global-dir.js";
+import { createHardhatRuntimeEnvironment } from "../../../../src/internal/hre-intialization.js";
 
 let hre: HardhatRuntimeEnvironment;
 let globalCacheDir: string;
